move to utils for ease of testing
